Refactor code to smaller, more concise footprint
